+++ app/Filament/Resources/CaptchaChallengeResource/Pages/ListCaptchaChallenges.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\CaptchaChallengeResource\Pages;
+
+use App\Filament\Resources\CaptchaChallengeResource;
+use Filament\Resources\Pages\ListRecords;
+
+class ListCaptchaChallenges extends ListRecords
+{
+    protected static string $resource = CaptchaChallengeResource::class;
+}